refactor: extract daemon artifact client
